Addressed type warnings
