+++ README.md
@@ -118,7 +118,9 @@ create a custom handler inside `@/src/factories/action-factory` to avoid code du
 Requires you to register a client id or access token in the developer portal of the third party
 storage provider. Currently, there is support for [Dropbox](https://www.dropbox.com/developers/apps), [Google Drive](https://console.cloud.google.com/) and S3 Buckets.
 
-You can enable each of these integrations by providing the required key values in the `.env` file.
+You can enable each of these integrations by providing the required key values for your configuration
+by creating a local `env.local`-file which will contain your custom files. You can create this
+file using the contents of `.env`-file.
 
 ## Project setup
